transactions is uint64_t.

+++ libdariadb/storage/chunk.h
@@ -34,7 +34,7 @@ struct ChunkHeader {
   size_t size;
   uint32_t crc;
 
-  uint32_t transaction; 
+  uint64_t transaction;
   bool commit : 1;
 };
 #pragma pack(pop)
